added auth

+++ graphql/resolvers/survey.js
@@ -14,7 +14,10 @@ export const surveyResolvers = {
               startup: 1,
               video: 1,
               feedbacks: 1,
+              questions: 1,
               fillers: 1,
+              eta: 1,
+              reward: 1,
             },
           },
           {
@@ -32,7 +35,9 @@ export const surveyResolvers = {
             },
           },
         ]);
-        console.log(surveys);
+        for (let survey of surveys) {
+          survey.startup = survey.startup[0];
+        }
         return surveys;
       } catch (e) {
         throw new Error(e);
@@ -41,7 +46,7 @@ export const surveyResolvers = {
     survey: async (_, { id }) => {
       try {
         const surveys = await survey.aggregate([
-          { $match: { $_id: id } },
+          { $match: { _id: id } },
           {
             $project: {
               _id: 1,
@@ -52,6 +57,9 @@ export const surveyResolvers = {
               video: 1,
               feedbackQst: 1,
               fillers: 1,
+              questions: 1,
+              eta: 1,
+              reward: 1,
             },
           },
           {
@@ -69,6 +77,8 @@ export const surveyResolvers = {
             },
           },
         ]);
+        surveys[0].startup = surveys[0].startup[0];
+        console.log(surveys[0]);
         return surveys[0];
       } catch (e) {
         throw new Error(e);
@@ -88,6 +98,8 @@ export const surveyResolvers = {
               video: 1,
               feedbackQst: 1,
               fillers: 1,
+              eta: 1,
+              reward: 1,
             },
           },
           {
@@ -105,6 +117,7 @@ export const surveyResolvers = {
             },
           },
         ]);
+        survey[0].startup = survey[0].startup[0];
         return surveys;
       } catch (e) {
         throw new Error(e);
@@ -112,8 +125,10 @@ export const surveyResolvers = {
     },
   },
   Mutation: {
-    createSurvey: async (_, args) => {
+    createSurvey: async (_, args, { user }) => {
+      if (user?.type != "startup") throw new Error("Unauthorized");
       try {
+        console.log(args.survey.questions.length);
         let newSurvey = new survey({
           ...args.survey,
           feedbacks: {
@@ -163,7 +178,9 @@ export const surveyResolvers = {
         throw new Error(e);
       }
     },
-    updateSurvey: async (_, args) => {
+    updateSurvey: async (_, args, { user }) => {
+      if (user?.type != "startup") throw new Error("Unauthorized");
+
       try {
         let updatedSurvey = await survey
           .findByIdAndUpdate(args.id, args.survey, { new: true })
@@ -198,13 +215,28 @@ export const surveyResolvers = {
         throw new Error(e);
       }
     },
-    deleteSurvey: async (_, args) => {
+    deleteSurvey: async (_, args, { user }) => {
+      if (user?.type != "startup") throw new Error("Unauthorized");
+
       try {
         await survey.findByIdAndDelete(args.id).lean();
         return {
           success: true,
           message: "Survey deleted successfully",
         };
+      } catch (e) {
+        throw new Error(e);
+      }
+    },
+    addResponse: async (_, args, {user}) => {
+      if (user?.type != "startup") throw new Error("Unauthorized");
+      try {
+        let res = await survey.findById(args.surveyId);
+        if (res.startup != user.id) throw new Error("Unauthorized");
+        let question = res.questions.find((q) => q.question === args.question);
+        question.responses.push(args.response);
+        await res.save();
+        return res;
       } catch (e) {
         throw new Error(e);
       }
